Revert "Merge branch 'master' into ZhiyeChen"
This reverts commit 4e96a159329883c20215f4b79a512bf6fd2748f8, reversing
changes made to d017687b5b1af0fc0ec6535a606edf27202a9e1f.

+++ Knowledge_Is_Power/ThunderExplode.java
@@ -57,7 +57,7 @@ public class ThunderExplode extends Bullet
         }else{
             turn(fire_rotation);
         }
-        move((int)(move_speed));
+        move(move_speed);
         setRotation(0);
     }
     public void gifAnimator(){
